myshop/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.9
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-parler==2.3
django-rosetta==0.10.1
django-taggit==5.0.1
easy-thumbnails==2.8.5
idna==3.10
joblib==1.4.2
kombu==5.4.2
localflavor==1.9
Markdown==3.6
MarkupSafe==2.1.5
nltk==3.9.1
numpy==2.1.2
oauthlib==3.2.2
pandas==2.2.3
parler==1.0.1
pillow==10.3.0
polib==1.2.0
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
redis==5.0.4
regex==2024.9.11
requests==2.31.0
requests-oauthlib==2.0.0
rosetta==0.3
scipy==1.14.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.4
sqlparse==0.5.1
stripe==11.1.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
